requirements.txt
discord
openai
python-dotenv
discord-py-interactions